+++ component_catalog/forms.py
@@ -624,8 +624,11 @@ class BaseAddToProductForm(
         super().__init__(user, *args, **kwargs)
 
         product_field = self.fields["product"]
-        perms = ["view_product", "change_product"]
-        product_field.queryset = Product.objects.get_queryset(user, perms=perms)
+        product_field.queryset = Product.objects.get_queryset(
+            user=user,
+            perms=["view_product", "change_product"],
+            exclude_locked=True,
+        )
 
         if relation_instance:
             help_text = f'"{relation_instance}" will be assigned to the selected product.'
@@ -709,7 +712,9 @@ class AddToProductAdminForm(forms.Form):
         self.relation_model = relation_model
         self.dataspace = request.user.dataspace
         self.fields["product"].queryset = Product.objects.get_queryset(
-            request.user, perms=["view_product", "change_product"]
+            user=request.user,
+            perms=["view_product", "change_product"],
+            exclude_locked=True,
         )
 
     def get_selected_objects(self):